Revert "Revert "refactor: move getAction to separate file for bundling""
This reverts commit 40deaeda1dd350b27fa3da994a7c37005ae7a187.
